requirements.txt
langchain>=0.2.16
langchain-core>=0.2.38
langchain-openai>=0.1.20
langchain-community>=0.2.16
langchain-text-splitters==0.2.4
duckduckgo-search
requests
beautifulsoup4
python-dotenv